+++ src/models/UserProfile.js
@@ -1,8 +1,0 @@
-/**
- * User profile factory class.
- */
-export class UserProfile {
-    name = "";
-    mail = "";
-    id = "";
-}