+++ CHANGELOG.md
@@ -9,13 +9,13 @@
 -   [DONE] Created Header.jsx component with proper styling.
 -   [DONE] Created Routes.js file to centralize Route table.
 
-### Changes Planned
+### Changes Alanned
 
 -   [TODO] Create new page Profile.
 
 ### 17/11/2022 - React Session # 9
 
-#### Changes Planned
+#### Changes Applied
 
 -   [DONE][0.3h] Create a contextAPI module to be used as a hook to other components.
 -   [WIP][0.1h] Consume the ContextAPI in a component
@@ -23,12 +23,74 @@
 -   [TODO][0.3h] Querying and storing data retrieved from an API.
 -   [TODO][1.0h] Create new page Profile.
 
-### 22/11/2022 - React Review # 10
+### 22/11/2022 - React Session # 10
 
-#### Changes applied
+#### Changes Applied
 
 -   [DONE] Created AppLayout component to add another Layout for contextAPI.
 -   [DONE] Worked on modifications on HomePage to display data.
 -   [DONE] Worked on Header logic to display profile.name.
 -   [DONE] Created UserProfile file & class.
 -   [DONE] Modified storage/context/index file to add new functions.
+
+### 29/11/2022 - React Session # 11
+
+#### Changes Planned
+
+-   [DONE] Removed MiniContextProvider form App.js
+-   [DONE] Modified AppLayout in order to consume new store.
+-   [DONE] Modified Header in order to consume new store.
+-   [DONE] Created useAppState hook.
+-   [DONE] Modified StateService service to use new store.
+
+### 01/12/2022 - React Session # 12
+
+#### Changes Planned
+
+-   [WIP][0.1h] Consume the State Layer in a component via hook
+-   [WIP][0.3h] Display the data stored in the app state hook.
+-   [TODO][0.3h] Redux - Querying and storing data retrieved from an API
+
+#### Changes Applied
+
+-   [DONE] Created AppLayout component to add another Layout for contextAPI.
+-   [DONE] Worked on modifications on HomePage to display data.
+-   [DONE] Modified UserProfile file to store "isAdmin".
+-   [DONE] Created UserProfileReducer file.
+-   [DONE] Created rootReducer file.
+-   [DONE] Created store file.
+-   [DONE] Modified StateService file to support new store.
+
+### 06/12/2022 - React Session # 13
+
+### Changes Applied
+
+-   [DONE][0.3h] Redux - Querying and storing fake user data - Fix any sync-up errors
+-   [TODO][0.5h] Redux - Querying and storing data retrieved from an API
+
+### 13/12/2022 - React Session # 14
+
+#### Changes Applied
+
+-   [DONE] Created AppAuthentication component to wrap App component and verify authentication before rendering it.
+-   [DONE] Removed all the auth logic from AppLayout component.
+-   [DONE] Created AuthService file.
+-   [DONE] Removed async from get functions in UserService.
+
+### 28/12/2022 - React Session # 15
+
+#### Changes Applied
+
+-   [DONE][0.3h] Comprehend, discuss and map necessary contract details.
+-   [DONE][0.3h] Created interfaces necessary to shape the state/api/service objects by means of contracts
+    -   IUserProfile.
+    -   IBoredAPIService.
+    -   IBoredAPIActivities.
+    -   IBoredActivity.
+-   [DONE][0.1h] Updated the User Profile reducers to include dispatch Payload types.
+-   [DONE][0.3h] Created BoredAPI service class to consume the third-party service.
+-   [DONE][0.1h] Created a function to call the created service on Homepage by means of a button (untested)
+
+### Future Work
+
+-   [TODO][1.0h] Create new page Profile.